requirements.txt
pandas
numpy
pyyaml
tabulate
pytest
